+++ packages/core/lib/clock.ts
@@ -1,7 +1,7 @@
 const NANOSECONDS_IN_MILLISECONDS = 1_000_000
 
 export function millisecondsToNanoseconds (milliseconds: number): number {
-  return milliseconds * NANOSECONDS_IN_MILLISECONDS
+  return Math.round(milliseconds * NANOSECONDS_IN_MILLISECONDS)
 }
 
 export interface Clock {