+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - (react-native) Added `net.host.connection.type` span attribute [#334](https://github.com/bugsnag/bugsnag-js-performance/pull/334)
+- (react-native) Added `net.host.connection.subtype` span attribute [#360](https://github.com/bugsnag/bugsnag-js-performance/pull/360)
 
 ### Fixed